gemini generate content with new skd, new model
