+++ community/gemini/.env.Example
