+++ community/gemini/.gitignore
